Move pytest dir to be unittest
