+++ tests/neuro_san/client/test_hocons.py
@@ -13,7 +13,7 @@ from unittest import TestCase
 
 from parameterized import parameterized
 
-from tests.framework.pytest.dynamic_hocon_unit_tests import DynamicHoconUnitTests
+from tests.framework.unittest.dynamic_hocon_unit_tests import DynamicHoconUnitTests
 
 
 class TestHocons(TestCase):
